rename(ui-design-tool): rename ui-models dir to api
