+++ packages/ui-design-tool/components/SelectionOverlay/useRenderUtils.tsx
@@ -1,6 +1,6 @@
+import { Layer } from '@/api/Layer/model';
+import { UIRecord } from '@/api/UIRecord/model';
 import { UIRecordRect } from '@/types/Geometry';
-import { Layer } from '@/ui-models/Layer/model';
-import { UIRecord } from '@/ui-models/UIRecord/model';
 import { useCallback } from 'react';
 import { useContextForInteraction } from '../Workspace/Workspace.context';
 import * as styles from './SelectionOverlay.css';
