+++ packages/ui-design-tool/components/TransformOverlay/types.ts
@@ -1,6 +1,6 @@
+import { UIRecord } from '@/api/UIRecord/model';
 import { UIRecordRect } from '@/types/Geometry';
 import { UIRecordKey } from '@/types/Identifier';
-import { UIRecord } from '@/ui-models/UIRecord/model';
 import ReactTypes from '@pigyuma/react-utility-types';
 
 export const HandlePlacement = {
