+++ packages/ui-design-tool/index.ts
@@ -1,11 +1,11 @@
 /** @todo Design token 모델 내보내기 */
 
-export * from './ui-models/Artboard/model';
-export * from './ui-models/Canvas/model';
-export * from './ui-models/Layer/model';
-export * from './ui-models/ShapeLayer/model';
-export * from './ui-models/TextLayer/model';
-export * from './ui-models/UIRecord/model';
+export * from './api/Artboard/model';
+export * from './api/Canvas/model';
+export * from './api/Layer/model';
+export * from './api/ShapeLayer/model';
+export * from './api/TextLayer/model';
+export * from './api/UIRecord/model';
 
 export { WorkspaceComponent as Workspace } from './components/Workspace/Workspace';
 export type { WorkspaceProps, WorkspaceRef } from './components/Workspace/types';
